+++ screens/DonutChart.tsx
@@ -1,9 +1,10 @@
 import React from 'react';
 import { View, StyleSheet, Dimensions, Text } from 'react-native';
 import { PieChart } from 'react-native-chart-kit';
+import styles from '../styles/donutStyles';
 
 const DonutChart : React.FC = () =>{
-  const data = [
+  const data = [ // 指定する要素は決まっているっぽいが、要素の追加は可能
     { name: "A", population: 35, color: "tomato", legendFontColor: "#7F7F7F", legendFontSize: 15 },
     { name: "B", population: 40, color: "orange", legendFontColor: "#7F7F7F", legendFontSize: 15 },
     { name: "C", population: 25, color: "gold", legendFontColor: "#7F7F7F", legendFontSize: 15 },
@@ -15,17 +16,17 @@ const DonutChart : React.FC = () =>{
           data={data}
           width={400}
           height={200}
-          chartConfig={{
+          chartConfig={{ // グラフの全体的なスタイル設定を行うオブジェクト
             backgroundColor: "#1cc910",
-            backgroundGradientFrom: "#eff3ff",
-            backgroundGradientTo: "#efefef",
+            backgroundGradientFrom: "#000",
+            backgroundGradientTo: "#000",
             color: (opacity = 1) => `rgba(0, 0, 0, ${opacity})`,
           }}
-          accessor={"population"}
-          backgroundColor={"transparent"}
+          accessor={"population"} // data配列から、円グラフの割合を描く要素を指定する
+          backgroundColor={"transparent"} // transparentを指定することで、背景を透明に設定
           paddingLeft={"100"}
-          center={[0, 0]} // グラフの中心を整える
-          absolute // 値を中央に表示する
+          center={[0, 0]} // グラフの中心を整える（左上を始点としている）
+          absolute // セグメント値（内訳みたいな説明が書かれたやつ）を中央に表示する
         />
         {/* ドーナツ中央部分 */}
         <View style={styles.centerText}>
@@ -36,51 +37,5 @@ const DonutChart : React.FC = () =>{
     </View>
   );
 };
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: "center",
-    alignItems: "center", // 画面の中心に配置
-  },
-  chartWrapper: {
-    position: "relative", // ドーナツ中央部分を重ねるため
-    justifyContent: "center",
-    alignItems: "center", // グラフを中心に配置
-    width: 200, // コンポーネント全体の横幅
-    height: 200, // コンポーネント全体の縦幅
-  },
-  centerText: {
-    position: "absolute",
-    width: 100, // 中央の円のサイズ
-    height: 100,
-    borderRadius: 50, // 完全な円形にする
-    backgroundColor: "white", // 背景色と同じ色
-    justifyContent: "center",
-    alignItems: "center",
-
-  },
-  centerTextTitle: {
-    fontSize: 16,
-    fontWeight: "bold",
-  },
-  centerTextValue: {
-    fontSize: 14,
-    color: "#7F7F7F",
-  },
-  donutHole: {
-    position: "absolute",
-    width: 100, // 中央の円のサイズ
-    height: 100,
-    borderRadius: 50, // 完全な円形にする
-    backgroundColor: "white", // 背景色と同じ色
-    justifyContent: "center",
-    alignItems: "center",
-  },
-  donutText: {
-    fontSize: 20,
-    fontWeight: "bold",
-  },
-});
 
 export default DonutChart;
